Move db utility functions out of uidEngine
