+++ src/ratesEngine.ts
@@ -2,10 +2,7 @@ import fetch from 'node-fetch'
 
 import { config } from './config'
 import { ExchangeRateReq } from './exchangeRateRouter'
-import { existsAsync, hgetallAsync, setAsync } from './uidEngine'
-import { createThrottledMessage } from './utils/createThrottledMessage'
-import { getEdgeAssetDoc } from './utils/dbUtils'
-import { slackPoster } from './utils/postToSlack'
+import { getEdgeAssetDoc, hgetallAsync, slackMessage } from './utils/dbUtils'
 import { logger, normalizeDate, snooze } from './utils/utils'
 
 const {
@@ -17,10 +14,6 @@ const {
 } = config
 
 const endPoint = `${ratesServerAddress}/v2/exchangeRates`
-const slackMessage = createThrottledMessage(
-  { set: setAsync, exists: existsAsync },
-  slackPoster
-)
 
 const LOOP_DELAY = 1000 * 30 // Delay 30 seconds
 const bridgeCurrency = DEFAULT_FIAT
